Move tests for AnalysesResultsManager into own file
These tests refer to the `AnalysesResultsManager` but at the moment they live in
`vscode-tests/no-workspace/remote-queries/remote-query-history.test.ts`

Let's give them their own file.
